Merge staging into arm1.1 (06/17/2020)

See merge request turbonomic/xl!1128

+++ com.vmturbo.sql.utils/src/main/java/com/vmturbo/sql/utils/MariaDBMySqlAdapter.java
@@ -15,7 +15,7 @@ import com.vmturbo.sql.utils.DbEndpoint.UnsupportedDialectException;
  */
 class MariaDBMySqlAdapter extends DbAdapter {
 
-    MariaDBMySqlAdapter(final DbEndpoint config) {
+    MariaDBMySqlAdapter(final DbEndpointConfig config) {
         super(config);
     }
 
@@ -26,6 +26,7 @@ class MariaDBMySqlAdapter extends DbAdapter {
         dataSource.setUrl(url);
         dataSource.setUser(user);
         dataSource.setPassword(password);
+        dataSource.setDatabaseName(config.getDbDatabaseName());
         return dataSource;
     }
 
@@ -33,21 +34,24 @@ class MariaDBMySqlAdapter extends DbAdapter {
     protected void createSchema() throws SQLException, UnsupportedDialectException, InterruptedException {
         try (Connection conn = getRootConnection(null)) {
             execute(conn, String.format("CREATE DATABASE IF NOT EXISTS %s",
-                    config.getDatabaseName()));
+                    config.getDbDatabaseName()));
         }
     }
 
     @Override
     protected void createNonRootUser() throws UnsupportedDialectException, SQLException, InterruptedException {
         try (Connection conn = getRootConnection(null)) {
-            dropUser(conn, config.getUserName());
+            dropUser(conn, config.getDbUserName());
             execute(conn, String.format("CREATE USER '%s'@'%%' IDENTIFIED BY '%s'",
-                    config.getUserName(), config.getPassword()));
+                    config.getDbUserName(), config.getDbPassword()));
         }
     }
 
     private void dropUser(final Connection conn, final String user) {
         // DROP USER IF NOT EXISTS is not supported in MySQL until v5.7, and breaks jenkins builds
+        if (!config.getDbDestructiveProvisioningEnabled()) {
+            return;
+        }
         try {
             execute(conn, String.format("DROP USER '%s@%%'", user));
         } catch (SQLException e) {
@@ -58,9 +62,9 @@ class MariaDBMySqlAdapter extends DbAdapter {
     @Override
     protected void performNonRootGrants(DbEndpointAccess access) throws SQLException, UnsupportedDialectException, InterruptedException {
         String privileges = getPrivileges(access);
-        try (Connection conn = getRootConnection(config.getDatabaseName())) {
+        try (Connection conn = getRootConnection(config.getDbDatabaseName())) {
             execute(conn, String.format("GRANT %s ON `%s`.* TO '%s'@'%%'",
-                    privileges, config.getDatabaseName(), config.getUserName()));
+                    privileges, config.getDbDatabaseName(), config.getDbUserName()));
         }
     }
