+++ com.vmturbo.sql.utils/src/main/java/com/vmturbo/sql/utils/PostgresAdapter.java
@@ -3,9 +3,15 @@ package com.vmturbo.sql.utils;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+import java.time.temporal.ChronoUnit;
+import java.util.Map;
 
 import javax.sql.DataSource;
 
+import com.google.common.collect.ImmutableMap;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.postgresql.ds.PGSimpleDataSource;
 
 import com.vmturbo.sql.utils.DbEndpoint.DbEndpointAccess;
@@ -14,34 +20,54 @@ import com.vmturbo.sql.utils.DbEndpoint.UnsupportedDialectException;
 /**
  * {@link DbAdapter} implementation for PostgreSQL endpoints.
  */
-class PostgresAdapter extends DbAdapter {
+public class PostgresAdapter extends DbAdapter {
 
-    PostgresAdapter(final DbEndpoint config) {
+    private static final Logger logger = LogManager.getLogger();
+
+    /**
+     * Supported time units for retention policy, mapping from time unit to sql query.
+     */
+    private static final Map<ChronoUnit, String> SUPPORTED_TIME_UNIT_FOR_RETENTION_POLICY =
+            ImmutableMap.of(
+                    ChronoUnit.YEARS, "years",
+                    ChronoUnit.MONTHS, "months",
+                    ChronoUnit.WEEKS, "weeks",
+                    ChronoUnit.DAYS, "days",
+                    ChronoUnit.HOURS, "hours"
+            );
+
+    PostgresAdapter(final DbEndpointConfig config) {
         super(config);
     }
 
     @Override
-    DataSource getDataSource(String url, String user, String password) throws InterruptedException {
+    DataSource getDataSource(String url, String user, String password) {
         final PGSimpleDataSource dataSource = new PGSimpleDataSource();
         dataSource.setUrl(url);
         dataSource.setUser(user);
         dataSource.setPassword(password);
-        dataSource.setCurrentSchema(config.getSchemaName());
+        // The flyway migrator (version 4.x) does not quote postgres schemas by default.
+        // This means that multi-tenant schema names (which contain "-", an illegal character
+        // in Postgres) do not get migrated properly. Add the quotes manually.
+        dataSource.setCurrentSchema("\"" + config.getDbSchemaName() + "\"");
         return dataSource;
     }
 
     @Override
     protected void createNonRootUser() throws SQLException, UnsupportedDialectException, InterruptedException {
         try (Connection conn = getRootConnection(null)) {
-            dropUser(conn, config.getUserName());
+            dropUser(conn, config.getDbUserName());
             execute(conn, String.format("CREATE USER \"%s\" WITH PASSWORD '%s'",
-                    config.getUserName(), config.getPassword()));
+                    config.getDbUserName(), config.getDbPassword()));
             execute(conn, String.format("ALTER ROLE \"%s\" SET search_path TO \"%s\"",
-                    config.getUserName(), config.getSchemaName()));
+                    config.getDbUserName(), config.getDbSchemaName()));
         }
     }
 
     private void dropUser(final Connection conn, final String user) throws SQLException {
+        if (!config.getDbDestructiveProvisioningEnabled()) {
+            return;
+        }
         try {
             execute(conn, String.format("DROP USER IF EXISTS \"%s\"", user));
         } catch (SQLException e) {
@@ -67,38 +93,40 @@ class PostgresAdapter extends DbAdapter {
     }
 
     private void performRWGrants() throws UnsupportedDialectException, SQLException, InterruptedException {
-        try (Connection conn = getRootConnection(config.getDatabaseName())) {
+        try (Connection conn = getRootConnection(config.getDbDatabaseName())) {
             execute(conn, String.format("GRANT ALL PRIVILEGES ON SCHEMA \"%s\" TO \"%s\"",
-                    config.getSchemaName(), config.getUserName()));
+                    config.getDbSchemaName(), config.getDbUserName()));
         }
     }
 
     private void performROGrants() throws SQLException, UnsupportedDialectException, InterruptedException {
-        try (Connection conn = getRootConnection(config.getDatabaseName())) {
+        try (Connection conn = getRootConnection(config.getDbDatabaseName())) {
             execute(conn, String.format("GRANT CONNECT ON DATABASE \"%s\" TO \"%s\"",
-                    config.getDatabaseName(), config.getUserName()));
+                    config.getDbDatabaseName(), config.getDbUserName()));
             execute(conn, String.format("GRANT USAGE ON SCHEMA \"%s\" TO \"%s\"",
-                    config.getSchemaName(), config.getUserName()));
+                    config.getDbSchemaName(), config.getDbUserName()));
             execute(conn, String.format("GRANT SELECT ON ALL TABLES IN SCHEMA \"%s\" TO \"%s\"",
-                    config.getSchemaName(), config.getUserName()));
+                    config.getDbSchemaName(), config.getDbUserName()));
             execute(conn, String.format("ALTER DEFAULT PRIVILEGES IN SCHEMA \"%s\" "
                             + "GRANT SELECT ON TABLES TO \"%s\"",
-                    config.getSchemaName(), config.getUserName()));
+                    config.getDbSchemaName(), config.getDbUserName()));
         }
     }
 
     @Override
     protected void createSchema() throws SQLException, UnsupportedDialectException, InterruptedException {
         try (Connection conn = getRootConnection(null)) {
-            if (!databaseExists(conn, config.getDatabaseName())) {
-                execute(conn, String.format("CREATE DATABASE \"%s\"", config.getDatabaseName()));
-                try (Connection dbConn = getRootConnection(config.getDatabaseName())) {
-                    execute(dbConn, "DROP SCHEMA public CASCADE");
+            if (!databaseExists(conn, config.getDbDatabaseName())) {
+                execute(conn, String.format("CREATE DATABASE \"%s\"", config.getDbDatabaseName()));
+                try (Connection dbConn = getRootConnection(config.getDbDatabaseName())) {
+                    if (config.getDbDestructiveProvisioningEnabled()) {
+                        execute(dbConn, "DROP SCHEMA public CASCADE");
+                    }
                 }
             }
         }
-        try (Connection conn = getRootConnection(config.getDatabaseName())) {
-            execute(conn, String.format("CREATE SCHEMA IF NOT EXISTS \"%s\"", config.getSchemaName()));
+        try (Connection conn = getRootConnection(config.getDbDatabaseName())) {
+            execute(conn, String.format("CREATE SCHEMA IF NOT EXISTS \"%s\"", config.getDbSchemaName()));
             setupTimescaleDb(conn);
         }
     }
@@ -113,11 +141,10 @@ class PostgresAdapter extends DbAdapter {
      *
      * @param conn db connection currently connected to the target database
      * @throws SQLException if there's a problem adding the extension
-     * @throws InterruptedException if interrupted
      */
-    protected void setupTimescaleDb(Connection conn) throws SQLException, InterruptedException {
-        execute(conn, String.format("CREATE EXTENSION IF NOT EXISTS timescaledb SCHEMA %s",
-                config.getSchemaName()));
+    protected void setupTimescaleDb(Connection conn) throws SQLException {
+        execute(conn, String.format("CREATE EXTENSION IF NOT EXISTS timescaledb SCHEMA \"%s\"",
+                config.getDbSchemaName()));
     }
 
     private boolean databaseExists(final Connection conn, final String databaseName) throws SQLException {
@@ -125,5 +152,47 @@ class PostgresAdapter extends DbAdapter {
                 String.format("SELECT * FROM pg_catalog.pg_database WHERE datname = '%s'",
                         databaseName));
         return results.next();
+    }
+
+    @Override
+    public void setupRetentionPolicy(String table, ChronoUnit timeUnit, int retentionPeriod)
+            throws UnsupportedDialectException, InterruptedException, SQLException {
+        // sanity check
+        if (retentionPeriod <= 0) {
+            logger.error("Invalid retention period provided: {}", retentionPeriod);
+            return;
+        }
+
+        final String timeUnitSql = SUPPORTED_TIME_UNIT_FOR_RETENTION_POLICY.get(timeUnit);
+        if (timeUnitSql == null) {
+            logger.error("Unsupported time unit {}", timeUnit);
+            return;
+        }
+
+        try (Connection conn = getNonRootConnection()) {
+            // do it in a transaction
+            conn.setAutoCommit(false);
+            // first drop previous policy if it exists
+            conn.createStatement().execute(String.format(
+                    "SELECT remove_drop_chunks_policy('%s', if_exists => true)", table));
+            // create new retention policy and get its background job id
+            ResultSet resultSet = conn.createStatement().executeQuery(String.format(
+                    "SELECT add_drop_chunks_policy('%s', INTERVAL '%d %s', "
+                            + "cascade_to_materializations => FALSE)", table, retentionPeriod, timeUnitSql));
+            if (!resultSet.next()) {
+                logger.error("Unable to create add_drop_chunks_policy for table \"{}\" with period \"{} {}\"",
+                        table, retentionPeriod, timeUnit);
+                return;
+            }
+
+            final int jobId = resultSet.getInt("add_drop_chunks_policy");
+            // set the drop_chunks background job to run every day, starting from next midnight
+            conn.createStatement().execute(String.format("SELECT alter_job_schedule(%d, "
+                    + "schedule_interval => INTERVAL '1 days', "
+                    + "next_start => date_trunc('DAY', now()) + INTERVAL '1 day');", jobId));
+            conn.commit();
+            logger.info("Created retention policy for table \"{}\" with period \"{} {}\"", table,
+                    retentionPeriod, timeUnit);
+        }
     }
 }
